Simplify the code flow.
Loops only have thread-affinity while running. This means, we can
*run* a loop within any thread [in the Executor].

So: instead of creating loops per-thread, we create them
per-connection and allow that loop to run within any available thread.
Knowing that __init__() is a synchronous call, we can also move much
of the blocking connection-setup into the constructor.

With the above changes, we can also toss the per-thread initializer
and the mapping of thread/loop.