General improvements, assertions, and THREAD_* arguments.
The MESSAGEID and HEADERS arguments have been used to construct a
threaded series of emails. These parameters have now been deprecated
in favor of THREAD_START and THREAD_KEY, for clarity. An appropriate
DeprecationWarning has been inserted.

This module still needs to work with Python 2.7 for id.a.o
(specifically: the sscommon.py module). Add a note/future that all
arguments should be provided as keywords, rather than positionally.

Added a few assertions on argument expectations.

We now use port 587 exclusively. The module used to default to
(maybe?) port 25 or maybe 465. It is unclear, but both ports are
incorrect.